+++ assets/javascripts/kitten/components/details/info.js
@@ -1,0 +1,23 @@
+export const info = {
+  text: `
+    # Details
+    &nbsp;
+
+    ## Import
+    ~~~js
+    import { Details } from '@kisskissbankbank/kitten/src/components/details'
+    ~~~
+
+    ## Usage
+    &nbsp;
+
+    #### Default
+    ~~~js
+    <Details summaryRender={ ({ isOpen }) => (…) }>
+      <p>My content</p>
+    </Details>
+    ~~~
+  `,
+  header: false,
+  propTables: false,
+}